Chatbot/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0

# Vector database
chromadb>=0.4.0
pinecone-client>=2.2.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
gradio>=3.40.0
streamlit>=1.25.0

# AWS SDK
boto3>=1.28.0
sagemaker>=2.180.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.65.0

# Optional: Local LLM support with Ollama
ollama>=0.1.7
requests>=2.31.0
